refactor: alternating encoder threads in case a frame can't be encoded within a frame interval (however the render thread and gpu can perform their work with in a frame interval), use two alternating encoder threads. threads as well as swapchain backbuffers are labeled as blue and green (e.g. this frame is blue, then the next will be green, the name is simply from how xcode instrucument metal system trace shows workloads of different frame). will there be a third one? you'd better figure out why encoding is that slow!

+++ src/d3d11/mtld11_resource_dynamic.cpp
@@ -108,7 +108,7 @@ private:
           },
           [](SIMPLE_RESIDENCY_TRACKER **ppTracker) {
             D3D11_ASSERT(0 && "todo: prepare dynamic buffer srv view");
-            return ArgumentData(0uL, 0);
+            return ArgumentData((uint64_t)0uL, 0);
           },
           [u = this->resource.ptr()](IMTLNotifiedBindable *_this) {
             u->RemoveObserver(_this);
